change float to double

+++ src/main/java/co/nvqa/common/core/model/order/Order.java
@@ -781,7 +781,7 @@ public class Order extends DataEntity<Order> implements Serializable {
     private String itemDescription;
     private Boolean isDangerousGood;
     private Integer quantity;
-    private Float unitValue;
+    private Double unitValue;
   }
 
   @Getter
